+++ ispra-lod-infrastructure/land/land.py
@@ -10,6 +10,7 @@ from typing import Dict, Callable
 import re
 from kg_loader import KnowledgeGraphLoader
 import pandas as pd
+from utils import label_en, label_it
 
 
 UNIT_OF_MEASURES = None
@@ -147,26 +148,6 @@ def place_type(istat, field):
         return None
 
     return "%s"%(type)
-
-
-def label_it(dset):
-    dset = str(dset)
-    if (dset == "soilc"):
-        return str("consumo del suolo")
-        #return TermUtils.irify("consumo del suolo")
-    elif (dset == "urban"):
-        return str("aree urbane")
-    else:
-        return None
-
-def label_en(dset):
-    dset = str(dset)
-    if (dset == "soilc"):
-        return str("soil consumption")
-    elif (dset == "urban"):
-        return str("urban areas")
-    else:
-        return None
 
 
 class LandTriplifier(Triplifier):
